src/backend/requirements_phase5.txt
# Phase 5 Authentication Requirements
# Add these to your main requirements.txt or install separately

# Database ORM and migrations
sqlalchemy>=2.0.0
alembic>=1.13.0

# PostgreSQL adapter (use if deploying with PostgreSQL)
# psycopg2-binary>=2.9.0

# JWT and password security
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4

# Pydantic settings
pydantic-settings>=2.0.0

# Optional: Redis for caching (for leaderboards, sessions)
# redis>=5.0.0

# Optional: Celery for background tasks (for emails, etc.)
# celery>=5.3.0

# Optional: Email sending
# python-multipart>=0.0.6
# aiosmtplib>=3.0.0

# Testing
pytest>=7.4.0
pytest-asyncio>=0.21.0
httpx>=0.25.0
